requirements.txt
streamlit>=1.31.0
faster-whisper==0.10.0
transformers==4.38.1
torch==2.2.0
python-dotenv>=1.0.0
websockets>=11.0.3
whitenoise==6.6.0
sentry-sdk==1.39.1 